+++ src/app/(components)/forms/CreateUserForm.jsx
@@ -17,7 +17,7 @@ const UserForm = () => {
       ...prevState,
       [name]: value,
     }));
-    console.log(JSON.stringify(formData));
+    // console.log(JSON.stringify(formData));
   };
 
   const handleSubmit = async (e) => {