+++ src/app/(components)/navigation/NavUser.jsx
@@ -6,7 +6,6 @@ import { usePathname } from "next/navigation"; // usePathname is a hook now impo
 
 const NavUser = () => {
   const pathname = usePathname();
-  console.log(pathname);
   return (
     <nav className={styles.navbar}>
       <div className={styles.leftLinks}>