+++ src/app/api/Timesheets/route.js
@@ -5,44 +5,50 @@ export async function POST(req) {
   try {
     const body = await req.json();
     const timesheetData = body.formData;
-    console.log(JSON.stringify(timesheetData));
 
     // Confirm all fields have been filled
     if (
       !timesheetData?.userEmail ||
-      !timesheetData?.entries ||
+      !timesheetData?.clientName ||
+      !timesheetData?.projectName ||
+      !timesheetData?.taskDescription ||
       !timesheetData?.date
     ) {
       return NextResponse.json({ message: "Missing fields" }, { status: 400 });
     }
 
-    await Timesheet.create(timesheetData);
-    return NextResponse.json(
-      { message: "Timesheet Created." },
-      { status: 201 },
-    );
+    // Check for duplicate entries
+    try {
+      await Timesheet.create(timesheetData);
+      return NextResponse.json(
+        { message: "Timesheet Created." },
+        { status: 201 },
+      );
+    } catch (error) {
+      // Check if the error is due to duplicate key violation
+      if (error.code === 11000) {
+        return NextResponse.json(
+          { message: "Duplicate entry found." },
+          { status: 400 },
+        );
+      }
+      throw error; // Throw error if it's not due to duplicate key violation
+    }
   } catch (error) {
     console.log(error);
-    if (error.code === 11000) {
-      // Duplicate key error
-      return NextResponse.json(
-        { message: "Duplicate timesheet entry" },
-        { status: 409 },
-      );
-    } else {
-      // Other errors
-      return NextResponse.json(
-        { message: "An error occurred while saving the timesheet" },
-        { status: 500 },
-      );
-    }
+    return NextResponse.json(
+      {
+        message:
+          "An error occurred while saving the timesheet. Please try again later.",
+      },
+      { status: 500 },
+    );
   }
 }
 
 export async function GET() {
   try {
     const timesheets = await Timesheet.find();
-    console.log(timesheets);
     return NextResponse.json({ timesheets }, { status: 201 });
   } catch (error) {
     console.log(error);
@@ -67,21 +73,17 @@ export async function PATCH(req) {
       );
     }
 
-    // Update taskDescription and time if they exist in the updatedFields
-    if (updatedFields.taskDescription) {
-      timesheet.entries.forEach((entry) => {
-        if (entry._id.toString() === updatedFields.entryId) {
-          entry.taskDescription = updatedFields.taskDescription;
-        }
-      });
+    if (updatedFields.clientName) {
+      timesheet.clientName = updatedFields.clientName;
     }
-
+    if (updatedFields.projectName) {
+      timesheet.projectName = updatedFields.projectName;
+    }
+    if (updatedFields.taskDescription) {
+      timesheet.taskDescription = updatedFields.taskDescription;
+    }
     if (updatedFields.time) {
-      timesheet.entries.forEach((entry) => {
-        if (entry._id.toString() === updatedFields.entryId) {
-          entry.time = updatedFields.time;
-        }
-      });
+      timesheet.time = updatedFields.time;
     }
 
     await timesheet.save();
@@ -131,21 +133,3 @@ export async function DELETE(req) {
     );
   }
 }
-
-/*
-Find way to filter data at read level
-export async function GET(req) {
-  try {
-    const {name} = req.query;
-
-    console.log(name)
-    console.log(userEmail)
-    const timesheets = await Timesheet.find({userEmail: "manager_test@bast.com"});
-    console.log(timesheets);
-    return NextResponse.json({ timesheets }, { status: 201 });
-  } catch (error) {
-    console.log(error);
-    return NextResponse.json({ message: "Error", error }, { status: 500 });
-  }
-}
-*/